Put __heap_base__ at end of internal RAM
...and not at the end of all allocated memory, which includes the sdram.
Normally, this is not a problem, unless the core firmware allocates
memory in the sdram. Normally all sdram is dedicated to the xpatch,
but if we want to utilize a memory allocator which uses sdram (as
we will in the AEC branch, because we need large amounts of memory),
then we will be using the sdram in the core firmware.