Bump BUFSIZE from 16 to 128 bytes
... in order to minimize the overhead when the Speex AEC is running.
The docs recommend using a period size of about 20ms, or 1024 samples
at 48 kHz. In that light, the default period size in the Axoloti of 16
samples is rather short, especially considering that the AEC runs FFTs
on the period size.

Moving beyond 128 samples causes the Axoloti core firmware to run out
of memory (more specifically .SRAM2).